[CLI]: Add --skip-unchanged flag for index command
In order to provide the same flexibility as the API, expose the
`SkipUnchanged` flag contained in `IndexOptions`. Although the indexing
process is changing rapidly, and so re-indexing all photos might the
right way to go for now, this flag will be very useful, once indexing is
stable.

For more info refer to [0]

[0]: https://github.com/photoprism/photoprism/issues/187

Signed-off-by: Linos Giannopoulos <14967949+linosgian@users.noreply.github.com>