Fix example configuration for Apache 2.4 (#573)
* Fix example configuration for Apache 2.4

The tutorial on serverlab.ca contains some errors. The official docs are accurate and more helpful. Added reminder for the benefit of people mindlessly copying and pasting.

* Portability fix for scripts/build.sh

As per the manual, the `-p` flag to `uname` is not portable and results in `unknown` in many cases, for instance, when compiling on a Raspberry Pi 4.