Update convert.go
When I upgraded my phone to the Google Pixel 7, I noticed that the videos it takes have an extra video stream in them for their compressed HVC1 implementation. This means that all software transcoding of that format would fail, as it was transcoding the wrong video stream. This alteration fixed the issue for me, and will likely avoid video files with multiple video streams causing issues in the future.